requirements.txt
anyio==3.6.2
autopep8==2.0.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
coverage==5.5
deprecation==2.1.0
eyed3==0.9.7
filetype==1.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.4
mysql-connector-python==8.0.31
packaging==22.0
Pillow==9.3.0
protobuf==3.20.1
pycodestyle==2.10.0
python-dotenv==0.21.0
python-telegram-bot==20.0b0
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.0
soupsieve==2.3.2.post1
toml==0.10.2
tomli==2.0.1
urllib3==1.26.13
